Updated User model, test and Gemfile
Still have some more things to work on with the user but I implemented some regex and uniqueness in the User model and the test checks out just fine. Updated Gemfile for spring having a bug of having to stop spring when trying to use the rails console when you have already opened and closed it.