Initial User commit
* Revert "Begun TDD"

This reverts commit e02cb8849caa69f3fdb095d268602fc686495dfc.

* Added User controller and model

Added User controller with `rails` command and model based on my user. I pulled the feature branch to the master branch because the 'wyatt' role was not available.

* User test and User app in the feature branch

I got a User test set and some User validation set as well. Passwords, regex and more validation are next.

* Updated User model, test and Gemfile

Still have some more things to work on with the user but I implemented some regex and uniqueness in the User model and the test checks out just fine. Updated Gemfile for spring having a bug of having to stop spring when trying to use the rails console when you have already opened and closed it.

* Refined user model and user test

I got the test working on the notion that the email has to be lowercase and added thaet to the user model

* User model has a password and has minimum length

The user model now has a password and a minimum length. Working on this, I had to get rid of everything in the bin/ directory and restore with
simple rake command. All tests passed as far as I am aware.